Refactor rest auth
