+++ harbor/rest/auth/password_reset.py
@@ -1,0 +1,72 @@
+'''This module contains all password reset related routes'''
+
+from fastapi import APIRouter, Depends, BackgroundTasks
+from pydantic import BaseModel, EmailStr
+from starlette.responses import JSONResponse
+from starlette.status import HTTP_401_UNAUTHORIZED
+
+from harbor.domain.common import ObjectIdStr, StrongPasswordStr, Message
+from harbor.repository.base import RepoDict, get_repos
+from harbor.use_cases.auth import (
+    reset_password_exec as uc_user_reset_pw_exec,
+    reset_password_req as uc_user_reset_pw_req,
+)
+
+
+router = APIRouter()
+
+
+class FormRequestPasswordReset(BaseModel):
+    '''Form to request a password reset'''
+    email: EmailStr
+
+
+@router.post("/login/request-password-reset/", response_model=Message)
+async def request_password_reset(form: FormRequestPasswordReset,
+                                 background_tasks: BackgroundTasks,
+                                 repos: RepoDict = Depends(get_repos)):
+    '''User requests a password reset'''
+    uc = uc_user_reset_pw_req.RequestPasswordResetUseCase(
+        user_repo=repos['user'],
+        vt_repo=repos['verif_token'],
+        background_tasks=background_tasks,
+    )
+    uc_req = uc_user_reset_pw_req.RequestPasswordResetRequest(
+        email=form.email
+    )
+    await uc.execute(uc_req)
+    return {'msg': 'Verification mail sent, if email is linked to an existing user'}
+
+
+class PasswordResetBody(BaseModel):
+    '''POST model to reset password'''
+    user_id: ObjectIdStr
+    token: str
+    password: StrongPasswordStr
+
+
+@router.post("/login/password-reset/", response_model=Message)
+async def password_reset(body: PasswordResetBody,
+                         repos: RepoDict = Depends(get_repos)):
+    '''Verifies password reset token and sets new password'''
+    uc = uc_user_reset_pw_exec.ExecResetPasswordUseCase(
+        user_repo=repos['user'],
+        vt_repo=repos['verif_token'],
+    )
+    uc_req = uc_user_reset_pw_exec.ExecResetPasswordRequest(
+        **body
+    )
+
+    try:
+        result = uc.execute(uc_req)
+        if result == uc_user_reset_pw_exec.ExecResetPasswordResponse.UPDATED:
+            return {'msg': 'Password is updated'}
+
+        if result == uc_user_reset_pw_exec.ExecResetPasswordResponse.UPDATED_AND_VERIFIED:
+            return {'msg': 'Password is updated and account is verified'}
+
+    except uc_user_reset_pw_exec.InvalidTokenError:
+        return JSONResponse(
+            status_code=HTTP_401_UNAUTHORIZED,
+            content={'msg': 'Provided token is not valid'}
+        )
